% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erdos_c.R
\name{erdos_c}
\alias{erdos_c}
\title{Undirected Erdos-Renyi random graph (C implementation)}
\usage{
erdos_c(n, p)
}
\arguments{
\item{n}{Number of vertices / nodes in the graph.}

\item{p}{probability of edge inclusion.}
}
\value{
adjacency list representation of the resultant graph.
}
\description{
\code{erdos_c} creates an adjacency list representation of an
undirected Erdos-Renyi random graph.
}
\details{
This function generates an undirected random graph of size \code{n}, with
the probability of any two nodes being connected, \code{p}.  Self and
repeated edges are not permitted.
}
\examples{
gr <- erdos_list(100, 0.1)
}
