% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erdos_r.R
\name{erdos_r}
\alias{erdos_r}
\title{Undirected Erdos-Renyi random graph (R implementation)}
\usage{
erdos_r(n, p)
}
\arguments{
\item{n}{number of vertices / nodes in the network.}

\item{p}{probability of edge between two nodes.}
}
\value{
adjacency list representation of the resultant network
}
\description{
\code{erdos_r} creates an adjacency list representation of an
undirected Erdos-Renyi random graph.
}
\details{
This function generates an undirected random graph of size \code{n}, with
the probability of any two nodes being connected, \code{p}.  Self and
repeated edges are not permitted.
}
\examples{
gr <- erdos_r(1000, 10)
}
