% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erdos_rcpp.R
\name{erdos_rcpp}
\alias{erdos_rcpp}
\title{Undirected Erdos-Renyi random graph (Rcpp implementation)}
\usage{
erdos_rcpp(n, p)
}
\arguments{
\item{n}{number of vertices / nodes in the network.}

\item{p}{probability of edge between two nodes.}
}
\value{
adjacency list representation of the resultant network
}
\description{
\code{erdos_rcpp} creates an adjacency list representation of an
undirected Erdos-Renyi random graph.
}
\details{
This function generates an undirected random graph of size \code{n}, with
the probability of any two nodes being connected, \code{p}.  Self and
repeated edges are not permitted.
}
