% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erdos_rust.R
\name{erdos_rust}
\alias{erdos_rust}
\title{Generate an erdos-renyi random graph}
\usage{
erdos_rust(n, p)
}
\arguments{
\item{n}{number of nodes}

\item{p}{probability of an edge between nodes}
}
\description{
Generate an erdos-renyi random graph
}
